requirements.txt
langchain-chroma~=0.2.2
langchain-community~=0.3.18
langchain-openai~=0.3.7
numpy~=1.26.4

# local
# python-dotenv~=1.0.1
